+++ apps/backoffice-v2/src/lib/blocks/components/CaseCallToActionLegacy/hooks/useCaseCallToActionLegacyLogic/useCaseCallToActionLegacyLogic.tsx
@@ -1,22 +1,25 @@
-import { useCallback, useMemo, useState } from 'react';
-import { useAuthenticatedUserQuery } from '../../../../../../domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
+import { useDocumentsByEntityIdsAndWorkflowIdQuery } from '@/domains/documents/hooks/queries/useDocumentsByEntityIdsAndWorkflowIdQuery/useDocumentsByEntityIdsAndWorkflowIdQuery';
 import { useWorkflowByIdQuery } from '@/domains/workflows/hooks/queries/useWorkflowByIdQuery/useWorkflowByIdQuery';
+import { useCaseState } from '@/pages/Entity/components/Case/hooks/useCaseState/useCaseState';
+import { useCallback, useMemo, useState } from 'react';
 import { useFilterId } from '../../../../../../common/hooks/useFilterId/useFilterId';
-import { TWorkflowById } from '../../../../../../domains/workflows/fetchers';
+import { useAuthenticatedUserQuery } from '../../../../../../domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
 import { useApproveCaseAndDocumentsMutation } from '../../../../../../domains/entities/hooks/mutations/useApproveCaseAndDocumentsMutation/useApproveCaseAndDocumentsMutation';
 import { useRevisionCaseAndDocumentsMutation } from '../../../../../../domains/entities/hooks/mutations/useRevisionCaseAndDocumentsMutation/useRevisionCaseAndDocumentsMutation';
-import { useCaseState } from '@/pages/Entity/components/Case/hooks/useCaseState/useCaseState';
+import { TWorkflowById } from '../../../../../../domains/workflows/fetchers';
 
 export const useCaseCallToActionLegacyLogic = ({
   parentWorkflowId,
   childWorkflowId,
   childWorkflowContextSchema,
+  isKYC,
 }: {
   parentWorkflowId: string;
   childWorkflowId: string;
   childWorkflowContextSchema: NonNullable<
     TWorkflowById['childWorkflows']
   >[number]['workflowDefinition']['contextSchema'];
+  isKYC: boolean;
 }) => {
   const filterId = useFilterId();
 
@@ -24,12 +27,13 @@ export const useCaseCallToActionLegacyLogic = ({
   const revisionReasons =
     childWorkflowContextSchema?.schema?.properties?.documents?.items?.properties?.decision?.properties?.revisionReason?.anyOf?.find(
       ({ enum: enum_ }) => !!enum_,
-    )?.enum as Array<string>;
+    )?.enum as string[];
 
   const noReasons = !revisionReasons?.length;
   const [reason, setReason] = useState(revisionReasons?.[0] ?? '');
   const [comment, setComment] = useState('');
   const reasonWithComment = comment ? `${reason} - ${comment}` : reason;
+
   // /State
 
   // Queries
@@ -39,31 +43,45 @@ export const useCaseCallToActionLegacyLogic = ({
     workflowId: parentWorkflowId,
     filterId,
   });
+
   const childWorkflow = parentWorkflow?.childWorkflows?.find(
     workflow => workflow.id === childWorkflowId,
   );
-  const nonIdentificationDocumentsIds = useMemo(() => {
+  const { data: documents } = useDocumentsByEntityIdsAndWorkflowIdQuery({
+    workflowId: parentWorkflowId,
+    entityIds: [childWorkflow?.context?.entity?.ballerineEntityId ?? ''],
+  });
+
+  const documentIds = useMemo(() => {
+    if (isKYC) {
+      return (
+        documents
+          ?.filter(document => document.type === 'identification_document')
+          ?.map(document => document.id) ?? []
+      );
+    }
+
     return (
       // 'identification_document' is exclusive to Veriff
-      childWorkflow?.context?.documents
+      documents
         ?.filter(document => document.type !== 'identification_document')
         ?.map(document => document.id) ?? []
     );
-  }, [childWorkflow?.context?.documents]);
+  }, [documents, isKYC]);
   // /Queries
 
   // Mutations
   const { mutate: mutateApproveCase, isLoading: isLoadingApproveCase } =
     useApproveCaseAndDocumentsMutation({
       workflowId: childWorkflowId,
-      ids: nonIdentificationDocumentsIds,
-      isDocumentsV2: !!parentWorkflow?.workflowDefinition?.config?.isDocumentsV2,
+      ids: documentIds,
+      isDocumentsV2: isKYC ? false : !!parentWorkflow?.workflowDefinition?.config?.isDocumentsV2,
     });
   const { mutate: mutateRevisionCase, isLoading: isLoadingRevisionCase } =
     useRevisionCaseAndDocumentsMutation({
       workflowId: childWorkflowId,
-      ids: nonIdentificationDocumentsIds,
-      isDocumentsV2: !!parentWorkflow?.workflowDefinition?.config?.isDocumentsV2,
+      ids: documentIds,
+      isDocumentsV2: isKYC ? false : !!parentWorkflow?.workflowDefinition?.config?.isDocumentsV2,
     });
   // /Mutations