+++ packages/ui/CHANGELOG.md
@@ -1,10 +1,19 @@
 # @ballerine/ui
 
+## 0.7.120
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.83
+
 ## 0.7.119
 
 ### Patch Changes
 
 - Bump
+- version bump
 
 ## 0.7.118