+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.105
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.105
+
+## 0.6.104
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.83
+  - @ballerine/workflow-core@0.6.104
+
 ## 0.6.103
 
 ### Patch Changes